[pf-862] Use csv helper
