+++ api/tests/integration/application/organizations/index_test.js
@@ -87,7 +87,7 @@ describe('Integration | Application | Organizations | Routes', () => {
   describe('POST /api/organizations/:id/invitations', () => {
 
     beforeEach(() => {
-      sinon.stub(securityController, 'checkUserIsOwnerInOrganization').callsFake((request, h) => h.response(true));
+      sinon.stub(securityController, 'checkUserIsOwnerInOrganizationOrHasRolePixMaster').callsFake((request, h) => h.response(true));
       sinon.stub(organisationController, 'sendInvitation').callsFake((request, h) => h.response().created());
 
       return server.register(route);
